refactor: RequestManager POP 형태로 변환
