+++ WeatherForecast/WeatherForecast/Controllers/ViewController.swift
@@ -10,9 +10,6 @@ class ViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view.
-        
-        
     }
 }
