+++ WeatherForecast/WeatherForecast/Network/HttpMethod.swift
@@ -1,0 +1,10 @@
+//
+//  HttpMethod.swift
+//  WeatherForecast
+//
+//  Created by Rarla on 2023/11/22.
+//
+
+enum HttpMethod {
+    case GET, POST, PUT, DELETE
+}
